Move `useUser` to standalone composable
